+++ src/napari_live_recording/processing_engine/image_filters/__init__.py
@@ -1,0 +1,11 @@
+from os.path import dirname, basename, isfile, join
+import glob
+
+
+class FilterFunction:
+    def __init__(self, parameters) -> None:
+        self.parameters = parameters
+        pass
+
+
+
