+++ src/napari_live_recording/processing_engine/image_filters/gauss.py
