+++ src/napari_live_recording/processing_engine/image_filters/other_gauss.py
