+++ src/napari_live_recording/processing_engine/testImage.jpg
